Add TLE classification and configuration records; refactor TLE creation methods to use new configuration structure
