+++ IO.Astrodynamics.Net/IO.Astrodynamics.Tests/Propagators/TLEPropagatorTests.cs
@@ -6,6 +6,7 @@ using IO.Astrodynamics.Body;
 using IO.Astrodynamics.Body.Spacecraft;
 using IO.Astrodynamics.Math;
 using IO.Astrodynamics.OrbitalParameters;
+using IO.Astrodynamics.OrbitalParameters.TLE;
 using IO.Astrodynamics.Propagator;
 using Xunit;
 using CelestialBody = IO.Astrodynamics.Body.CelestialBody;
