+++ IO.Astrodynamics.Net/IO.Astrodynamics/OrbitalParameters/OrbitalParameters.cs
@@ -4,8 +4,8 @@ using IO.Astrodynamics.Body;
 using IO.Astrodynamics.Coordinates;
 using IO.Astrodynamics.Frames;
 using IO.Astrodynamics.Math;
-using IO.Astrodynamics.SolarSystemObjects;
 using IO.Astrodynamics.TimeSystem;
+using IO.Astrodynamics.OrbitalParameters.TLE;
 
 namespace IO.Astrodynamics.OrbitalParameters;
 
@@ -32,7 +32,7 @@ public abstract class OrbitalParameters : IEquatable<OrbitalParameters>
     /// </summary>
     public Frame Frame { get; }
 
-    private TLEMeanElementsConverter _tleMeanElementsConverter;
+    private MeanElementsConverter _meanElementsConverter;
 
     //Data used for caching
     private Vector3? _eccentricVector;
@@ -80,7 +80,7 @@ public abstract class OrbitalParameters : IEquatable<OrbitalParameters>
         Observer = observer ?? throw new ArgumentNullException(nameof(observer));
         Epoch = epoch;
         Frame = frame ?? throw new ArgumentNullException(nameof(frame));
-        _tleMeanElementsConverter = new TLEMeanElementsConverter();
+        _meanElementsConverter = new MeanElementsConverter();
     }
 
 
@@ -975,42 +975,37 @@ public abstract class OrbitalParameters : IEquatable<OrbitalParameters>
     }
 
     /// <summary>
-    /// Converts the orbital parameters to a TLE (Two-Line Element) format.
-    /// This method fits the TLE from the current state vector and returns a TLE object.
+    /// Converts the current orbital parameters into a TLE (Two-Line Element) representation
+    /// using the provided configuration.
     /// </summary>
-    /// <param name="noradId">The NORAD ID of the satellite.</param>
-    /// <param name="name">The name of the satellite.</param>
-    /// <param name="cosparId">The COSPAR ID of the satellite.</param>
-    /// <param name="revolutionsAtEpoch">The number of revolutions at the epoch.</param>
-    /// <param name="classification">The classification of the satellite (default is 'U' for unclassified).</param>
-    /// <param name="firstDerivativeMeanMotion">
-    /// The first derivative of the mean motion (default is 0.0).
+    /// <param name="config">
+    /// A <see cref="Astrodynamics.OrbitalParameters.TLE.Configuration"/> object containing the necessary parameters for TLE generation,
+    /// such as NORAD ID, satellite name, COSPAR ID, and other optional settings.
     /// </param>
-    /// <param name="secondDerivativeMeanMotion">
-    /// The second derivative of the mean motion (default is 0.0).
-    /// </param>
-    /// <param name="bstarDragTerm">
-    /// The BSTAR drag term, representing atmospheric drag (default is 0.0001).
-    /// </param>
-    /// <param name="tol">
-    /// The tolerance for the fitting process (default is 1E-03).
-    /// </param>
-    /// <param name="maxIterations">
-    /// The maximum number of iterations for the fitting process (default is 15).
-    /// </param>
-    /// <param name="elementSetNumber">Element set number (default is 9999)</param>
     /// <returns>
-    /// A TLE object representing the orbital parameters in Two-Line Element format.
+    /// A <see cref="TLE"/> object representing the orbital parameters in TLE format.
     /// </returns>
-    public TLE ToTLE(ushort noradId, string name, string cosparId, ushort revolutionsAtEpoch, char classification = 'U',
-        double firstDerivativeMeanMotion = 0.0,
-        double secondDerivativeMeanMotion = 0.0, double bstarDragTerm = 0.0001, double tol = 1, ushort maxIterations = 15, ushort elementSetNumber = 9999)
+    /// <remarks>
+    /// This method uses the <see cref="MeanElementsConverter"/> to compute the mean orbital elements
+    /// from the current orbital parameters. The resulting TLE includes metadata such as the satellite's
+    /// classification, drag term, and epoch information.
+    /// </remarks>
+    /// <exception cref="ArgumentNullException">
+    /// Thrown if the <paramref name="config"/> parameter is null.
+    /// </exception>
+    public TLE.TLE ToTLE(TLE.Configuration config)
     {
-        var meanElements = _tleMeanElementsConverter.Convert(this, noradId, name, cosparId, revolutionsAtEpoch, bstarDragTerm, tol, maxIterations);
-        return TLE.Create(meanElements, name, noradId, cosparId, revolutionsAtEpoch, classification, bstarDragTerm, firstDerivativeMeanMotion, secondDerivativeMeanMotion,
-            elementSetNumber);
+        ArgumentNullException.ThrowIfNull(config);
+        var meanElements = _meanElementsConverter.Convert(this, config.NoradId, 
+            config.Name, config.CosparId, config.RevolutionsAtEpoch, 
+            config.BstarDragTerm, config.Tolerance, config.MaxIterations);
+    
+        return TLE.TLE.Create(meanElements, config.Name, config.NoradId, 
+            config.CosparId, config.RevolutionsAtEpoch, config.Classification, 
+            config.BstarDragTerm, config.FirstDerivativeMeanMotion, 
+            config.SecondDerivativeMeanMotion, config.ElementSetNumber);
     }
-
+    
     #region Operators
 
     public override bool Equals(object obj)
